\name{murmuratR-package}
\alias{murmuratR-package}
\alias{murmuratR}
\docType{package}
\title{
\packageTitle{murmuratR}
}
\description{
\packageDescription{murmuratR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{murmuratR}
\packageIndices{murmuratR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{murmuratR}

Maintainer: \packageMaintainer{murmuratR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
