% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_flock.R
\name{new_flock}
\alias{new_flock}
\title{Initialise a data frame of birds}
\usage{
new_flock(n)
}
\arguments{
\item{n}{The number of birds in the flock}
}
\value{
A new data frame
}
\description{
Initialise a data frame of birds
}
\examples{
  my_flock <- new_flock(1000)
}
